requirements.txt
torch>=1.10
torchvision
gradio==4.15.0
opencv-python
numpy
Pillow
timm
flask
